gnome/requirements.txt
flask
zappa
PyGithub
requests
ipaddress
PyYAML==3.12
pytest
pylint
python-dotenv
envparse
